Add an annotation to edited k8s resources
The annotation has a string value that is a serialized json object with
its keys being JSON pointers of fields in the annotated resource and
values being a list of PEM bundle indices that were edited (or "N/A" if
the field was not PEM-encoded).

For example:

```
metadata:
  annotations:
    ...
    recert-edited: '{"/spec/config/storage/files/15/contents/source":["1","2","6","5","3","0","4"],"/spec/config/storage/files/21/contents/source":["0"]}'
    ...
```

This annotation is purely informational and is not used programmatically
by recert itself. It's used to help troubleshooters notice this is not a
"natural" cryptographic object, but instead one that was manipulated by
recert.

Resolves https://issues.redhat.com/browse/MGMT-15538