Sync webhook-authenticator static secret files
The webhook authenticator secret has a kubeConfig field that is too
complicated to handle in recert. We could simply delete it and it will
be reconciled, but that's a bit too slow for us as it causes a
kube-apiserver rollout. To speed things up, we'll just "reconcile" it
ourselves by copying the kubeConfig contents from the kubeConfig file on
disk that we already processed with recert.

Also enable the preserve_order feature of serde_json to make sure field
order is maintained, because those differences also cause rollouts