Modify both `webhook-authenticator` and `webhook-authenticator-<latest_revision>`
Turns out both `webhook-authenticator` and
`webhook-authenticator-<latest_revision>` have to be the same. I
initially assumed the one without the suffix is simply the first
revision, so I deleted it, but it actually always equals (actually just
the data segment should be equal) to the latest revision, otherwise a
rollout is triggered. So we have to make sure both are updated.