Enable vendoring
Cachito doesn't do Rust, so we have to vendor all of our dependencies.

We're doing Vendoring using [`cargo vendor-filterer`](https://github.com/coreos/cargo-vendor-filterer) because
otherwise regular cargo pulls in a lot of heavy dependencies that are Windows only, which
will make the git repo very big for a tool that is used in Linux only.