Fix kubeconfig YAML serialization in etcd
After switching to search and replace for the kubeconfig field in the
respective controller-manager and kube-scheduler ConfigMaps, we
shouldn't serialize the edited kubeconfig field to YAML, as it includes
an additional '|' character. The latter doesn't lead to an error, but it
does lead to additional rollout triggers for these components.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>