Fix additional kube-apiserver config rollout
In OCP 4.15 the service-account-jwks-uri config arg of the
kube-apiserver contains a URI with the api label, instead of the api-int
one and the cluster domain. This change ensures that both api and
api-int labels are matched when replacing the cluster domain, in order
to keep compatibility with previous OCP versions.