Tuck some fields into structs
A recent addition of a lot of rename functionality made the code a bit
unreadable.

This commit tucks all rename-related fields into a struct called
`ClusterCustomizations` so passing them around is more neat.

Did the same for timings related code, and put it in a module, so it's
more out of sight from the main code flow.