Support editing PEM bundles lacking trailing newline
When a PEM bundle lacks a trailing newline,
`pem_bundle_replace_pem_at_index` would error when replacing the final
PEM in the bundle.

This happened because before replacing, the function ensures that it can
only see that exactly PEM once in the bundle, as a sanity check. It does
so by parsing it, encoding it, then searching for it.

The issue was that while encoding it, the resulting PEM string would have a
newline at the end of it, so while searching for it in the bundle, it
would not find any result - because the last PEM, as it appears in the
original bundle, doesn't have a newline that follows it.

The fix here would be to `trim_end` the encoded PEM before searching for
it.

But this also means that when finally replacing it, we need to make sure
that we're replacing it with the desired PEM, encoded without a newline
at the end, otherwise we would add a newline that wasn't there and
potentially cause rollouts.