Include error in recert summary
When recert runs into an error, it returns it as a `Result` from the
`main` function. This causes the Rust runtime to print the error trace.

However, the error does not get included in the recert summary files. In
the summary, you only see the logs, which makes it look like no error
occurred.

This commit makes it so that when recert runs into an error the error
will be included in summary files