Compute hash annotations for the {openshift,oauth}-apiserver
This change ensures that rollouts after recert for the following components are
skipped:

- openshift-apiserver/openshift-apiserver
- openshift-oauth-apiserver/oauth-apiserver

The goal is the reduction of the stabilization time of the Ingress and the
Authentication cluster operators.

Specifically, those rollouts are triggered by the difference in annotations on
the respective deployments. The latter include:

- a spec-hash (SHA256 of the deployment.spec) annotation
- multiple ConfigMap and Secret fnv32 hash annotations of objects that the
  respective cluster-operators are watching

In order to correctly compute the spec-hash annotations, we need to ensure that
whatever the respective cluster-operators compute the hash on, is what we use as
well. Because the operators include a mix of YAML and Go code for the specs they
compute the spec-hash on, we do the same with a hack script written in Go.
The latter:

- downloads the required deployment YAML manifests from the specific OCP
  release branch
- encodes them in JSON (mainly because JSON encoding between Go and Rust are
  quite different)
- adds some spec fields that can be hardcoded and leaves some templated spec
  fields to be filled in in Rust code (e.g. container image pull-specs,
  revisions, etc.).

The fnv32 hash annotations are also adjusted to follow the Go JSON encoding
implementation:

- secret data fields are lexicographically ordered
- when a field is a nil pointer it's encoded as the 'null'
  JSON type

Signed-off-by: Michail Resvanis <mresvani@redhat.com>