Add support for machine-network-cidr rename
# Background

The `cluster-config-v1` configmaps contain a copy of
`install-config.yaml` that was used for installation. In that
install-config, one of the fields is the machine-network CIDR.

The cluster-network-operator reads the machine-network field from the
`cluster-config-v1` configmaps while it tries to determine the final
value for `noProxy`, so that means we must ensure those configmaps have
the current machine network value.

# Current support

Currently recert supports replacing the entire install-config in those
configmaps. This is good for IBU, where we already have a cluster we can
take the entire install-config from.

# Issue

In IBI, we would have to fake such install-config for the fresh cluster.
This is inconvenient.

# Solution

Add a new `--machine-network-cidr` rename option to recert, which would
replace just the machine-network-cidr in the existing (seed)
install-config, rather than forcing the user to supply the entire
machine-config.