Reset `release-4.15` to `main`
We're not interested to continue backporting to 4.15, but we want 4.15
to have all the new features we work on. Since `main` likely didn't
introduce any regressions since `4.15` branched out of it, we prefer to
simply reset `release-4.15` to `main`.

This PR's diff was generated by simple running:

`git restore --source=main .`

In practice making 4.15 be the same as main.

No fancy merging / conflict resolution or anything