Fix type inference regression in indirect dependency
# Background

Recert uses the `simple_asn1` crate in order to calculate the SKID on
RSA keys

# Problem

The `simple_asn1` crate depends on the `time` crate, which recently
encountered a regression [1] due to [2] that caused the `simple_asn1`
crate to fail to compile with our locked version (0.3.30) of the `time`
crate.

# Solution

Updated the `time` crate to version 0.3.36, which contains a fix for the
regression. Re-ran vendoring.

[1] https://github.com/time-rs/time/issues/681
[2] https://github.com/rust-lang/rust/issues/125319