jwt: calculate kid from public key
# Background

Recert resigns JWTs found in the cluster. The kid field in the
Kubernetes JWT header is a Base64URL encoded hash of the public key used
to sign the JWT. The kid used by verifiers (e.g. kube-apiserver) to
quickly identify the key used to sign the JWT, instead of brute-forcing.

# Issue

Likely to solve issues.redhat.com/browse/OCPBUGS-49972

The current implementation uses the SHA256 hash of the private key to
calculate the kid. This is inconsistent with the Kubernetes
implementation [1] which uses the public key to calculate the kid.

In recent versions of Kubernetes, a kid mismatch leads to "unauthorized"
errors, which eventually leads to a collapse of the cluster.

# Solution

This commit changes the kid calculation to use the SHA256 hash of the
public key instead of the private key, so that it becomes more
consistent with Kubernetes

# Backwards compatibility

Seeds processed with older versions of recert that don't contain this
fix should still work if "rejuvenated" by newer versions of recert which
do contain this fix. This is because the new version will simply
overwrite the `kid` field with the correct value.

# References

[1] https://github.com/openshift/kubernetes/blob/6fdacf04117cef54a0babd0945e8ef87d0f9461d/pkg/serviceaccount/jwt.go#L92-L112

Signed-off-by: Omer Tuchfeld <omer@tuchfeld.dev>