Identify external certs by hash rather than just name
# Background

Recert looks at particular configmaps in the cluster to identify which
certs are external to the cluster and thus should not be regenerated, so
they are ignored from the rest of the recert process

# Issue

Sometimes people trust certs from other clusters that have the same name
as certs inside the cluster. This causes recert to skip internal certs
because it only compares the name.

# Solution

This commit changes the way we identify external certs to use the hash
of the DER bytes of the certificate, instead of just the name. This
method should be more robust and avoid this mistaken ignoring.

# Other changes

Will now use a new type called `ExternalCerts` to store the name and
hash of the certificates instead of a HashMap. Since it gets passed
around a lot its best to have it as a single type to make future changes
easier.