Parse CBOR into JSON
If we encounter CBOR data (identified by the self-describing tag), we
parse it and convert it internally to JSON.

For now, we don't support writing CBOR back, we just write JSON back,
which should work, but it means the cluster will lose its CBOR
representation.

In the next commits, we'll add the ability to write CBOR back as well.